+++ src/main/java/cn/taketoday/context/ConcurrentProperties.java
@@ -170,7 +170,7 @@ public class ConcurrentProperties extends Properties implements ConcurrentMap<Ob
 
 	@Override
 	protected void rehash() {
-
+		//
 	}
 
 	@Override